move isStdLib to internal pkg
